add seperate store for timeline
