+++ frontend/src/features/map_planning/store/MapStoreTypes.ts
@@ -317,15 +317,6 @@ export const UNTRACKED_DEFAULT_STATE: UntrackedMapState = {
     }),
     {} as UntrackedLayers,
   ),
-  timeLineEvents: {
-    daily: [],
-    monthly: [],
-    yearly: [],
-  },
-  timeLineVisibleYears: {
-    from: new Date().getFullYear() - 100,
-    to: new Date().getFullYear() + 100,
-  },
 };
 
 /**
@@ -392,36 +383,6 @@ export type TrackedPlantLayerState = {
    * The objects that have been loaded from the backend.
    */
   loadedObjects: PlantingDto[];
-};
-
-export type TimelineDailyEvent = {
-  key: number;
-  year: number;
-  month: number;
-  day: number;
-  added: number;
-  removed: number;
-};
-
-export type TimelineMonthlyEvent = {
-  key: number;
-  year: number;
-  month: number;
-  added: number;
-  removed: number;
-};
-
-export type TimelineYearlyEvent = {
-  key: number;
-  year: number;
-  added: number;
-  removed: number;
-};
-
-export type TimeLineEvents = {
-  daily: TimelineDailyEvent[];
-  monthly: TimelineMonthlyEvent[];
-  yearly: TimelineYearlyEvent[];
 };
 
 export type TrackedBaseLayerState = {
@@ -521,11 +482,6 @@ export type UntrackedMapState = {
   tooltipPosition: { x: number; y: number };
   bottomStatusPanelContent: React.ReactNode | null;
   layers: UntrackedLayers;
-  timeLineEvents: TimeLineEvents;
-  timeLineVisibleYears: {
-    from: number;
-    to: number;
-  };
 };
 
 /**
